refactor: Move the view actions helper to a separate package
